requirements.txt
impacket>=0.11.0
pytest>=8.0.0
ldap3>=2.9.0
pycryptodome>=3.15.0
requests>=2.25.0
neo4j>=5.0.0
dnspython>=2.0.0  # Optional: Enhanced DNS resolution via Domain Controller
bhopengraph>=1.1.0  # Official BloodHound OpenGraph library
